Fix NVP marking all routers as Error
DHC-1877

The NvpSynchronizer class in Neutron runs
a synchronization thread to sync nvp objects with
neutron objects. Since we don't use nvp's routers
the sync was failing making neutron showing all
the routers like if the were in Error state.
To fix this behaviour we override the two methods
responsible for the routers synchronization in the
NvpSynchronizer class to be a noop

Change-Id: I2d3a392d352474774824bc7cec579b2c9bbae472
Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>