Upgrade the nsx driver to Icehouse
DHC-2494

This commit copies all the interested methods from the original
nsx driver in neutron and applies a few changes we need to make
it work with akanda.
Note that the driver has been renamed nsx in neutron, it was
called nvp in havana.

Signed off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>

Change-Id: I0d8d4dbeeabd5717771e29245339362beecf0477