Add support for the ml2 plugin
Now that the ovs monolithic plugin has been deprecated we need
to switch to the ml2 plugin selecting ovs as mechanism driver.
All the methods we were overriding in the ovs plugin regarding
both L2 and L3 layers need now to be overriden in the two new
Ml2Plugin and L3RouterPlugin classes.

Change-Id: I08f904f2dc1f86139c44b8b208e57e5dd4ab552c
Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>